place desktop file in favour of a menu files
Switching from a *.menu file to a *.desktop file. This is also driven by
the CTTE decision in 2015/09 in bug report #741573 that also suggest to
switch from *.menu files to *.desktop.

The .desktop specification got some significant technical benefits over
the .menu description. There is no need to stay on the not easy to
maintain old *.menu files.

 https://lists.debian.org/debian-devel-announce/2015/09/msg00000.html