debian/copyright: adjust copyright information
The latest upstream changes infects some changes to the patch queue and
so some files have changed as they got removed. Also the change of the
inclusion of the additional tarballs for i18n, documentation and
libraries files involving some changes to the correct copyright
information.
Adding the CC-BY-SA-3.0 license as policy states that every license need
to be included inside debian/copyright.