debian/get-kicad.sh: modify script to get recent versions
Fully rewriting the script and renaming it to
create_kicad_upstream_tarballs.sh to get automated generated all
tarballs. Just call the script from top of the repository with the
needed version of kicad to get the four tarballs that will used by
git-buildpackage.

The script generate the following tarballs:

   kicad_<version>+dfsg<x>.orig.tar.xz
   kicad_<version>+dfsg<x>.orig-doc.tar.xz
   kicad_<version>+dfsg<x>.orig-i18n.tar.xz
   kicad_<version>+dfsg<x>.orig-libraries.tar.xz

The DFSG version is currently hardcoded an need to be tweaked by hand
if necessary. The script need a new option for creating tarballs with
flexible DFSG versions.

Note that the script is using some bashismus and needs the bash to work
properly! That should be fixed later.