rebuild patch queue from patch-queue branch
added patches:
fixes/Add-missing-dependency-github_plugin-pcbcommon.patch
fixes/Remove-unused-FindOpenSSL.cmake.patch
fixes/Replace-avhttp-with-libcurl.patch
fixes/Replace-the-avhttp-library-used-by-the-github-plugin-with.patch

Those patches adding support for using the libcurl library instead of
avhttp which is disabled as it would be needed to download the sources
while built.
Thanks to Simon Richter for collecting all the needed stuff to use
libcurl instead of avhttp and enables this backport.
avhttp got removed after the final release of version 4 of KiCad and
libcurl was added.