fix some buildflag hardening issues
We need to inject some more flags by concatenating the provided flags
from dpkg into the flags used by cmake.
CMake is ignoring CPPFLAGS but debhelper >= 9.20120417 is appending the
needed flags internally if we provide the prepared flags.
This commit fixes not all the existing hardening issues but most of
them.

See also
https://wiki.debian.org/Hardening?#Notes_for_packages_using_CMake