rework helper script for source tarball creation
In preparation for the next upcomming release 4.0.5 there are
improvements done to the script
debian/create_kicad_upstream_tarballs.sh.

The script was missing some checking for a valid download URL that was
resulting in empty source tarballs. Fixed by checking if the download
for the given version is possible. The script will now break if one of
the core source tarballs (main application, documentation or i18n) can't
be done.

Adding a alternative download for the KiCad main application searched in
GitHub if the release can't be found on Launchpad.

Adding the Connectors_TE-Connectivity.pretty library to the download
list for kicad-libraries, it's not empty anymore.