debian/rules: adjusting KICAD_FULL_VERSION
Upstream has changed the behaviour for creating the version string.
KICAD_VERSION is set intern statically, the user can add a string
KICAD_BRANCH_NAME (which we don't use) and a string KICAD_VERRION_EXTRA
if needed.
We use the latter to combining the Debian version behind the upstream
version. The debian version is created from the changelog entry.

By this the user will see the full Debian version in the UI of the KiCad
project manager and in every help menu.

Note that we need to patch the file CMakeModules/WriteVersionHeader.cmake
for doing this! Otherwise we would have a hyphen inside the version
string.