debian/rules: remove build-arch specific target
A desirable factorited debian/rules with full support for -indep and -arch
spedific ovrrides needs some more rework and can at least currently not
work. Reverting for now the target override_dh_auto_build-arch.

The package kicad-common needs files and folders from the arch specific
build if the kicad package. So we need to build currently also arch
specific things to get a sucessful binary-indep build.
To solve this we need to get rid of the kicad-common package which
itself has no big advantages from a current view on the things. And this
would need a rework of the control file and the kicad-common dh
sequencer helper files.