Revert "debian/control: decrease Architectures for arch packages"
This reverts commit 1835f2e4c21a568cd2aa7550c1337cebcd6b9cf4.

The referenced commit was the wrong approach to solve some not build
related issues. Changing the build architecture isn't useful in all,
wanna build is handling this correctly for packages that have
unsatisfied build dependencies. The reason for the initial issue was the
change of the Architecture field for kicad-common.

Thanks to Emelio Pozuelo monfort and Aurelien Jarno that helps to
understand the underlaying problem!