debian/control: move Maintainer to PKG Electronics Team
George and myself came to the conclusion it would be better to move the
content of the Maintainer field to the PKG Electronics Team.
Most of the other EDA tools are bundled in that team and because of that
the package would become a wider attention and also more channels for
communication and resolving some questions.

I asked also Richard A. Burton about his name in the Uploader field and
possible future work of Richard on KiCad. Richard agreed that his name
can and should be removed from that field, it's long ago he was active
working on packaging. So this commit also removes Richard from the
uploaders.

Richard, let me say a big thank you for maintaining KiCad for Debian in
the past!