kicad: remove a potentially RUNPATH from Pyhton2 lib
Upstream is building two libraries from the same source within pcbnew/
which are basically the same.
The library _pcbnew.so will become /usr/bin/_pcbnew.kiface, that is the
shared library for pcbnew. But it's also used as library for the Python
interface and moved finally to /usr/lib/python2.7/dist-packages.

The problem is a RUNPATH which is build into the original "mother"
library which is removed while moving the library over to _pcbnew.kiface
but not while creating the Python 2 library.

Using a global switch "-DCMAKE_SKIP_RPATH=TRUE" isn't helpful as this
removes *all* RPATH or RUNPATH variables from the DSOs which would turn
in a gab so the libraries that are needed from the KiCad binaries
wouldn't be found.
But removing any RUNPATH from usr/lib/python2.7/dist-packages/_pcbnew.so
ensures that kicad will be using the system default paths and let
ldconfig do the work for providing the path to the required libraries.

The issue of the not cleared RPATH is a upstream issue and some kind
confirmed, currently it's unknown if upstream will work on the issue.

See also https://wiki.debian.org/RpathIssue
         https://cmake.org/Wiki/CMake_RPATH_handling