kicad-libraries: adding new virtual package
The previous approach to hold all common used files in the package
kicad-common isn't working well now as e.g. the 3D models have been
grown dramatically and it's better to keep all these files in seperate
packages. To keep it usable for the users adding a virtual package
kicad-libraries that is pulling the typical used libraries for
footprints, symbols and templates.