autopkg: remove the only existing autopkgtest
The one autopkgtest we have used was created for testing the correct
existence of libraries within the new created package. The maintenance
of the various upstream libraries has been changed for the release
circle of KiCad 5 and we don't need the old autopkg check for now, so
remove this completely.