debian/rules: turn off support of SCRIPTING_WXPYTHON
Also related to the broken GTK+3 things in wxWidgets we also need to
disable the SCRIPTING_WXPYTHON feature, currently there is no wxpython
related library available with GTK+2 bindings. There is only
python-wxgtk3.0-dev which would bring some symbol clashing because this
brings a dependency on python-wxgtk3.0 and by this a depending on
libwxgtk3.0-gtk3-0v5.

Closes: #895801, #899215, #899216