kicad-doc: split docu build into two runs
While preparing the packaging of 5.1.0 it happen a classical build of
kicad-doc is failing due some internal cunfusioness within the CMake
magic and the Polish build of EEschema docu get mixed up with
Japanese.
This commit is working around this by split the build of the documentation
into two separate calls with two dedicated subfolders.